fix mock files
