+++ src/__mocks__/@/background/messenger/strict/api.ts
@@ -15,7 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { SanitizedIntegrationConfig } from "@/integrations/integrationTypes";
+import { RemoteResponse } from "@/types/contract";
+import { NetworkRequestConfig } from "@/types/networkTypes";
 import { RegistryId } from "@/types/registryTypes";
+import { Nullishable } from "@/utils/nullishUtils";
 import { getMethod, backgroundTarget as bg } from "webext-messenger";
 
 export * from "../../../../../background/messenger/strict/api";
@@ -29,4 +33,27 @@ export const registry = {
     );
   }),
   clear: getMethod("REGISTRY_CLEAR", bg),
+};
+
+export const services = {
+  locateAllForId: jest.fn().mockResolvedValue([]),
+  locate: jest
+    .fn()
+    .mockRejectedValue(new Error("Locate not implemented in mock")),
+  refresh: jest.fn(),
+  refreshLocal: getMethod("LOCATOR_REFRESH_LOCAL", bg),
+};
+
+// `getMethod` currently strips generics, so we must copy the function signature here
+export const performConfiguredRequestInBackground = getMethod(
+  "CONFIGURED_REQUEST",
+  bg,
+) as <TData>(
+  integrationConfig: Nullishable<SanitizedIntegrationConfig>,
+  requestConfig: NetworkRequestConfig,
+) => Promise<RemoteResponse<TData>>;
+
+export const dataStore = {
+  get: jest.fn().mockRejectedValue(new Error("Not implemented in mock")),
+  set: getMethod("SET_DATA_STORE", bg),
 };
